squash lint warnings

+++ MagickCore/colormap.c
@@ -373,7 +373,7 @@ MagickExport MagickBooleanType SortColormapByIntensity(Image *image,
       ssize_t
         i;
 
-      i=ConstrainColormapIndex(image,GetPixelIndex(image,q),exception);
+      i=ConstrainColormapIndex(image,(ssize_t) GetPixelIndex(image,q),exception);
       index=(Quantum) pixels[i];
       SetPixelIndex(image,index,q);
       SetPixelViaPixelInfo(image,image->colormap+(ssize_t) index,q);
